Replace interface tests with ParsePlanUseCase implementation
